app/requirements.txt
Django==1.10.6
djangorestframework==3.6.2
djangorestframework-expiring-authtoken==0.1.4
coreapi==2.3.0
markdown==2.6.8
pygments==2.2.0
requests==2.13.0
django-beanstalkd==1.0.1
